Added main menu page
New main menu page and styling has been added -- this template (the stuff surrounding the actual menu items) is going to be what we probably use for the whole shopping/menu experience up until the cart stage. TODO -- I fucked up and I think we should move the logo into the header along with the "Menu" title. I set up a grid to do that but haven't had a chance to do it yet, but that's what it's there for. I'll have to refactor a lot of the code to fix the layout, but it'll look way better once that's done.