Mobile menu is mostly done (but see push notes)
I got this to the point where the layout works. That being said, we still have some problems. I'm not sure how to fix them, but I feel like we're in a good enough place that we can submit what we have for the homework tonight. Here are the problems we still have that, at the time being, I'm not totally sure how to fix:

CLOSING MOBILE MENU:
Once it's opened, the hamburger (now an X) that's supposed to close the menu is totally gone. Not sure what's going on here, but I'll try and figure it out soon.

SUB-MENUS (hours and language):
These are totally buggy. The hours are pretty much fine, I just need to fix the z-layer issue so that the other text in the mobile menu isn't showing up on top of it. The language menu is more complicated; for whatever reason, it's totally stretched out to the point that nothing in it is actually visible. I think these can probably both be fixed; the issue is really just that I made these using only CSS and having absolutely no understanding of what I was actually doing. We can ask for some help here, perhaps.

CODE:
There's got to be a ton of redundant code in here suddenly, especially since I was repurposing code that I didn't originally write. I tried to fish most of it out, but there's likely still a good deal of it in there. We can go through and try to clean it up soon.