Incrementers now work
Redid incrementing framework. We don't really have an actual back-end right now (maybe eventually we will? I kind of want to learn to use PHP/SQL and that could make this stuff easier) so it doesn't really do anything, but it's nice that it's interactive now. Next task is gonna be to keep people from being able to order negative amounts of items. That shouldn't be too hard to fix, I just need to google how if/else statements work in JavaScript.